+++ components/DisputeResolver.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { View, ActivityIndicator } from 'react-native';
-import Button from '@/ui/primitives/Button';
+import { Button } from '@/ui';
 import OrderService from '@/services/orders';
 
 interface Props {
