+++ components/admin/AdminShell.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { View, Text } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { useTheme } from '@/ui/ThemeProvider';
-import Divider from '@/ui/primitives/Divider';
+import { Divider } from '@/ui';
 
 export default function AdminShell({ title, children, actions }: {
   title: string;
