+++ jest.config.js
@@ -21,6 +21,7 @@ module.exports = {
     '^@providers/(.*)$': '<rootDir>/src/providers/$1',
     '^@/features/(.*)$': '<rootDir>/src/features/$1',
     '^@/shared/(.*)$': '<rootDir>/src/shared/$1',
+    '^@/ui$': '<rootDir>/src/ui/index',
     '^@/ui/(.*)$': '<rootDir>/src/ui/$1',
     '^@/layout/(.*)$': '<rootDir>/src/layout/$1',
     '^@/services$': '<rootDir>/src/services/index',
