+++ src/features/auth/AuthContext.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext, ReactNode, useEffect, useState } from 'react';
 import { Alert } from 'react-native';
-import { Spinner } from '@/ui/primitives';
+import { Spinner } from '@/ui';
 import DatabaseService from '@/services/database';
 import { User } from '@/types';
 import { errorLog } from '@/utils/logger';
