+++ src/features/payments/components/MoonPayModal.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Modal, View, StyleSheet, Platform } from 'react-native';
 import { WebView } from 'react-native-webview';
-import { Spinner } from '@/ui/primitives';
+import { Spinner } from '@/ui';
 import { useTheme } from '@/ui/ThemeProvider';
 import { useAppInfo } from '@/contexts/AppInfoContext';
