+++ src/shared/ui/EmptyState.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { View, StyleSheet } from 'react-native';
-import Text from '@/ui/primitives/Text';
+import { Text, Button } from '@/ui';
 import type { LucideProps } from 'lucide-react-native';
 import { useTheme } from '@/ui/ThemeProvider';
-import Button from '@/ui/primitives/Button';
 
 type IconComponent = React.ComponentType<LucideProps>;
